chore: 리뉴얼 이전 파일들 정리
