--- web-app/css/post_list.css
@@ -1,195 +1,0 @@
-@charset "utf-8";
-@import url(reset.css);
-
-
-.post-list {
-    width: 90%;
-    height: 100%;
-    box-sizing: border-box;
-    padding: 25px;
-    margin: 60px auto;
-    border: 1px solid var(--main-color);
-    border-radius: 10px;
-    box-shadow: 0 1px 3px 0 var(--main-color);
-}
-
-.post-list-title {
-    font-size: 24px;
-}
-
-.post-list-bar {
-    display: block;
-    border: 1px solid #aaaaaa;
-    margin: 15px 0;
-    width: 35%;
-}
-
-.post-list-items {
-    position: relative;
-    display: grid;
-    grid-template-columns: 5fr 1fr;
-    grid-template-rows: 2fr;
-    box-sizing: border-box;
-    padding: 15px;
-    margin-bottom: 30px;
-    cursor: pointer;
-
-    /*color: rgba(223, 190, 106, 0.7);*/
-    transition: all 0.5s ease-out;
-    /*background: linear-gradient(270deg, rgba(223, 190, 106, 0.8), rgba(146, 111, 52, 0.8), rgba(34, 34, 34, 0), rgba(34, 34, 34, 0));*/
-    background: linear-gradient(270deg, rgba(95, 128, 128, 0.8), rgba(47, 79, 79, 0.8), rgba(34, 34, 34, 0), rgba(34, 34, 34, 0));
-    background-position: 1% 50%;
-    background-size: 300% 300%;
-    text-decoration: none;
-    border: 1px solid rgba(47, 79, 79, 0.3);
-}
-.post-list-items:last-child{
-    margin-bottom: 0;
-}
-
-.post-list-items:hover {
-    background-position: 99% 50%;
-    color: #ffffff;
-    border: 1px solid rgba(47, 79, 79, 0);
-}
-
-/*.post-list-items:after {
-    grid-column: 1/4;
-    content: '';
-    width: 100%;
-    margin: 0 auto;
-    border: 1px solid var(--main-color);
-}*/
-
-.post-list-items-title {
-    display: inline-block;
-    margin: 10px 0;
-    font-size: 20px;
-}
-
-.post-list-items-date {
-    margin: 10px 0;
-    font-size: 16px;
-    text-align: right;
-}
-
-.post-list-items-text {
-    grid-column: 1/3;
-    margin-bottom: 30px;
-    font-size: 14px;
-    overflow: hidden;
-}
-
-/*
-.post-list-items:hover > .post-list-items-barTop,
-.post-list-items:active > .post-list-items-barTop {
-    animation: barTop 1.5s ease-in-out 0s forwards;
-}
-
-.post-list-items:hover > .post-list-items-barLeft,
-.post-list-items:active > .post-list-items-barLeft {
-    animation: barLeft 1.5s ease-in-out 0s forwards;
-}
-
-.post-list-items:hover > .post-list-items-barRight,
-.post-list-items:active > .post-list-items-barRight {
-    animation: barRight 1.5s ease-in-out 0s forwards;
-}
-
-.post-list-items:hover > .post-list-items-barBottom,
-.post-list-items:active > .post-list-items-barBottom {
-    animation: barBottom 1.5s ease-in-out 0s forwards;
-}
-
-.post-list-items-bar {
-    position: absolute;
-    display: block;
-    background: var(--main-color);
-}
-
-.post-list-items-barTop {
-    top: 0;
-    left: 0;
-    height: 3px;
-    width: 0;
-}
-
-.post-list-items-barLeft {
-    top: 0;
-    left: 0;
-    width: 3px;
-    height: 0;
-}
-
-.post-list-items-barRight {
-    top: 0;
-    right: 0;
-    width: 3px;
-    height: 0;
-}
-
-.post-list-items-barBottom {
-    bottom: 0;
-    !*left: 0;*!
-    right: 0;
-    width: 0;
-    height: 3px;
-}
-
-@keyframes barTop {
-    0% {
-        width: 0;
-    }
-    !*20%{width: 100%;}*!
-    !*40%{width: 100%;}*!
-    !*60%{width: 100%;}*!
-    !*80%{width: 100%;}*!
-    !*100%{width: 0;}*!
-    100% {
-        width: 100%
-    }
-}
-
-@keyframes barLeft {
-    0% {
-        height: 0;
-    }
-    !*20%{height: 100%;}*!
-    !*40%{height: 100%;}*!
-    !*60%{height: 100%;}*!
-    !*80%{height: 100%;}*!
-    !*100%{height: 0;}*!
-    100% {
-        height: 100%
-    }
-
-}
-
-@keyframes barRight {
-    0% {
-        height: 0;
-    }
-    !*20%{height: 0;}*!
-    !*40%{height: 100%;}*!
-    !*60%{height: 100%;}*!
-    !*80%{height: 0;}*!
-    !*100%{height: 0;}*!
-    100% {
-        height: 100%
-    }
-
-}
-
-@keyframes barBottom {
-    0% {
-        width: 0;
-    }
-    !*20%{width: 0;}*!
-    !*40%{width: 100%;}*!
-    !*60%{width: 100%;}*!
-    !*80%{width: 0;}*!
-    !*100%{width: 0;}*!
-    100% {
-        width: 100%
-    }
-}*/
