--- web-app/css/searchResult.css
@@ -1,89 +1,0 @@
-@charset "utf-8";
-@import url(reset.css);
-@import url(header.css);
-@import url(aside.css);
-
-#wrap {
-    position: relative;
-    width: 100%;
-    height: 100%;
-    max-width: 1903px;
-}
-
-/* main 부분 */
-#main {
-    position: relative;
-    z-index: 1;
-    height: 100%;
-    box-sizing: border-box;
-    padding: 70px 100px;
-    margin-top: 60px;
-    transition: all 1s;
-    overflow: auto;
-}
-#main.reduce {
-    width: 88%;
-    left: 12%;
-}
-
-#main.increase {
-    width: 100%;
-    left: 0;
-    padding: 70px 150px;
-}
-
-.main-visual {
-    position: relative;
-    margin: 0 auto;
-    width: 100%;
-    height: 100%;
-}
-
-.searchResult{
-    width: 90%;
-    height: 100%;
-    box-sizing: border-box;
-    padding: 20px;
-    margin: 0 auto;
-    border: 2px solid #999999;
-    border-radius: 20px;
-}
-.searchResult-title{
-    font-size: 20px;
-}
-.searchResult-title-searchWord{
-    font-size: 22px;
-}
-.searchResult-bar{
-    display: block;
-    width: 100%;
-    margin: 10px 0;
-    border: 1px solid #999999;
-}
-.searchResult-lists{
-
-}
-.searchResult-lists-contents{
-    width: 96%;
-    margin: 0 auto;
-}
-.searchResult-lists-contents::after{
-    display: block;
-    width: 100%;
-    margin: 10px 0;
-    border: 1px solid #999999;
-    content: "";
-}
-.searchResult-lists-contents-title{
-    display: inline-block;
-    margin: 10px 0;
-    font-size: 18px;
-}
-.searchResult-lists-contents-date{
-    margin-bottom: 10px;
-    font-size: 16px;
-}
-.searchResult-lists-contents-text{
-    margin-bottom: 30px;
-    font-size: 14px;
-}
