+++ web-app/old-files/blogSettings.css
@@ -1,0 +1,206 @@
+@charset "utf-8";
+@import url(reset.css);
+@import url(header.css);
+@import url(aside.css);
+
+#wrap {
+    position: relative;
+    width: 100%;
+    height: 100%;
+    max-width: 1903px;
+}
+
+/* main 부분 */
+#main {
+    position: relative;
+    z-index: 1;
+    height: 100%;
+    box-sizing: border-box;
+    padding: 70px 100px;
+    margin-top: 60px;
+    transition: all 1s;
+    overflow: auto;
+}
+
+#main.reduce {
+    width: 88%;
+    left: 12%;
+}
+
+#main.increase {
+    width: 100%;
+    left: 0%;
+    padding: 70px 150px;
+}
+
+#main-visual {
+    position: relative;
+    margin: 0 auto;
+    width: 100%;
+    height: 100%;
+}
+
+
+/* 블로그 설정 영역 */
+#myBlog{
+    border: 1px solid #000;
+    padding: 20px 0;
+    box-sizing: border-box;
+    width: 70%;
+    margin: 0 auto;
+}
+#myBlog-title{
+    font-size: 26px;
+    font-weight: bold;
+    margin: 15px 0;
+    text-indent: 10%;
+}
+.myBlog-bar{
+    display: block;
+    border: 1px solid #aaaaaa;
+    margin: 25px auto;
+    width: 85%;
+}
+.myBlog-container{
+    display: block;
+    width: 80%;
+    margin: 15px auto;
+}
+.myBlog-container-titles{
+    font-size: 20px;
+    margin-bottom: 15px;
+}
+.myBlog-members-labels{
+    display: block;
+    margin-bottom: 10px;
+    font-size: 18px;
+    text-indent: 10px;
+}
+
+#myBlog-name-subject,
+#myBlog-members-inviteLink{
+    display: inline-block;
+    border: 1px solid #000;
+    width: 86.5%;
+    height: 36px;
+    font-size: 20px;
+    margin: 0 0 10px 10px;
+    text-indent: 10px;
+    box-sizing: border-box;
+    /*float: left;*/
+}
+.myBlog-name-description,
+.myBlog-members-inviteLinkDescription,
+.myBlog-membersList-Description{
+    display: block;
+    margin-bottom: 10px;
+    font-size: 15px;
+}
+#myBlog-members-inviteLinkButton{
+    display: inline-block;
+    width: 12%;
+    height: 36px;
+    border: 1px solid #000;
+    /*margin: 0 0 10px 10px;*/
+    float: right;
+    cursor: pointer;
+}
+#myBlog-members-inviteLinkButton:hover,
+#myBlog-members-inviteLinkButton:active{
+    background: skyblue;
+    color: #ffffff;
+}
+#myBlog-members-inviteLinkButton:after{
+    display: block;
+    content: "";
+    clear: both;
+}
+#myBlog-members-inviteLinkButton::after{
+    display: block;
+    content: "";
+    clear: both;
+}
+.myBlog-container{
+    display: block;
+    content: "";
+    clear: both;
+}
+#myBlog-membersList-addButton{
+    display: inline-block;
+    width: 12%;
+    height: 36px;
+    border: 1px solid #000;
+    margin: 0 0 10px 10px;
+    cursor: pointer;
+}
+#myBlog-membersList-addButton:hover,
+#myBlog-membersList-addButton:active{
+    background: skyblue;
+    color: #ffffff;
+}
+#myBlog-membersList-search{
+    display: inline-block;
+    /*width: 30%;*/
+    height: 36px;
+    border: 1px solid #000;
+    border-radius: 18px;
+    margin: 0 0 10px 10px;
+    vertical-align: top;
+    float: right;
+}
+.myBlog-membersList-searchButton{
+    font-size: 20px;
+    /*line-height: 2;*/
+    margin: 8px 0 0 10px;
+    cursor: pointer;
+}
+#myBlog-membersList-search > input{
+    /*border: 1px solid #000;*/
+    height: 33px;
+    vertical-align: top;
+    text-indent: 5px;
+}
+.myBlog-membersList-users-title{
+    font-size: 16px;
+}
+.myBlog-membersList-users-accessLevel{
+    font-size: 16px;
+    float: right;
+    margin-right: 50px;
+}
+
+.myBlog-membersList-users-list-image{
+    width: 50px;
+    float: left;
+}
+.myBlog-membersList-users-list-name{
+    display: block;
+    font-size: 16px;
+    text-indent: 20px;
+    width: 50%;
+}
+.myBlog-membersList-users-list-email{
+    display: block;
+    font-size: 15px;
+    text-indent: 20px;
+    width: 50%;
+
+}
+.myBlog-membersList-users-list-accessLevel{
+    float: right;
+    margin-top: -39px;
+    border: 1px solid #000;
+    font-size: 16px;
+    cursor: pointer;
+    padding: 5px 10px;
+}
+.myBlog-membersList-users-list-accessLevel:hover,
+.myBlog-membersList-users-list-accessLevel:active{
+    background: #cccccc;
+}
+.myBlog-bar-inside{
+    width: 100%;
+    display: block;
+    border: 1px solid #aaaaaa;
+    margin: 25px auto;
+}
