+++ web-app/old-files/dashboard_mainPage.css
@@ -1,0 +1,260 @@
+@charset "utf-8";
+@import url(reset.css);
+@import url(header.css);
+@import url(aside.css);
+
+#wrap {
+    position: relative;
+    width: 100%;
+    height: 100%;
+    max-width: 1903px;
+}
+
+/* main 부분 */
+#main {
+    position: relative;
+    z-index: 1;
+    height: 100%;
+    box-sizing: border-box;
+    padding: 70px 100px;
+    margin-top: 60px;
+    transition: all 1s;
+    overflow: auto;
+}
+#main.reduce {
+    width: 88%;
+    left: 12%;
+}
+
+#main.increase {
+    width: 100%;
+    left: 0;
+    padding: 70px 150px;
+}
+
+.main-visual {
+    position: relative;
+    margin: 0 auto;
+    width: 100%;
+    height: 100%;
+}
+
+.dashboard{
+    width: 90%;
+    height: 100%;
+    margin: 0 auto;
+}
+.dashboard-pinnedIssues-title{
+    font-size: 20px;
+
+}
+.dashboard-pinnedIssues-lists{
+    display: inline-block;
+    width: 100%;
+}
+.dashboard-pinnedIssues-lists > li{
+    float: left;
+    width: 32%;
+    height: 70px;
+    box-sizing: border-box;
+    padding-top: 10px;
+    margin-top: 10px;
+    margin-right: 2%;
+    border: 1px solid #000000;
+    border-radius: 6px;
+    text-indent: 5px;
+}
+
+.dashboard-pinnedIssues-list-title{
+    font-size: 18px;
+}
+.dashboard-pinnedIssues-list-title:hover,
+.dashboard-pinnedIssues-list-title:active {
+    font-weight: bold;
+}
+.dashboard-pinnedIssues-list-small{
+    text-indent: 10px;
+}
+
+.dashboard-filter{
+    display: inline-block;
+    position: relative;
+    width: 100%;
+    height: 40px;
+    margin-top: 15px;
+    border-radius: 6px;
+}
+.dashboard-filter-show{
+    height: 40px;
+    width: 90px;
+    border: 1px solid #000000;
+    cursor: pointer;
+}
+.dashboard-filter-show:hover,
+.dashboard-filter-show:active{
+    background: #cccccc;
+}
+.dashboard-filter-input{
+    position: absolute;
+    top: 0;
+    left: 90px;
+    width: 70%;
+    height: 100%;
+    box-sizing: border-box;
+    border: 1px solid #000;
+    border-left: none;
+}
+.dashboard-filter-input-icon{
+    position: absolute;
+    z-index: 10;
+    margin: 10px 0 0 10px;
+    font-size: 22px;
+    cursor: pointer;
+}
+.dashboard-filter-input-text{
+    position: absolute;
+    right: 0;
+    width: 94%;
+    height: 100%;
+    font-size: 18px;
+}
+
+.dashboard-newIssueButton{
+    position: absolute;
+    right: 0;
+    width: 100px;
+    height: 40px;
+    border: 1px solid #000000;
+    border-radius: 6px;
+    cursor: pointer;
+}
+.dashboard-newIssueButton:hover,
+.dashboard-newIssueButton:active{
+    background: lightgreen;
+}
+
+.dashboard-lists{
+    margin-top: 30px;
+}
+
+.dashboard-lists-caption,
+.dashboard-lists > li:first-child{
+    position: relative;
+    height: 50px;
+    border: 1px solid #000;
+}
+.dashboard-lists-caption-checkbox{
+    position: absolute;
+    top: 15px;
+    left: 15px;
+    width: 20px;
+    height: 20px;
+}
+.dashboard-lists-caption-opensIcon{
+    position: absolute;
+    top: 15px;
+    left: 60px;
+    font-size: 20px;
+}
+.dashboard-lists-caption-opens{
+    position: absolute;
+    top: 13px;
+    left: 85px;
+    font-size: 16px;
+}
+.dashboard-lists-caption-closesIcon{
+    position: absolute;
+    top: 15px;
+    left: 170px;
+    font-size: 20px;
+}
+.dashboard-lists-caption-closes{
+    position: absolute;
+    top: 13px;
+    left: 195px;
+    font-size: 16px;
+}
+.dashboard-lists-caption-assignees{
+    position: absolute;
+    top: 15px;
+    right: 120px;
+    font-size: 16px;
+}
+.dashboard-lists-caption-sorts{
+    position: absolute;
+    top: 15px;
+    right: 30px;
+    font-size: 16px;
+}
+
+.dashboard-lists > li{
+    position: relative;
+    height: 60px;
+    border: 1px solid #000;
+}
+
+.dashboard-lists-checkbox{
+    position: absolute;
+    top: 20px;
+    left: 15px;
+    width: 20px;
+    height: 20px;
+}
+
+.dashboard-lists-opensIcon{
+    position: absolute;
+    top: 15px;
+    left: 60px;
+    font-size: 18px;
+}
+
+.dashboard-lists-subject{
+    position: absolute;
+    top: 8px;
+    left: 85px;
+    font-size: 18px;
+}
+.dashboard-lists-subject:hover,
+.dashboard-lists-subject:active{
+    font-weight: bold;
+}
+.dashboard-lists-subject-notice{
+    padding: 0 7px;
+    margin-left: 10px;
+    background: yellow;
+    border-radius: 15px;
+    font-size: 16px;
+    font-weight: bold;
+}
+
+.dashboard-lists-small{
+    position: absolute;
+    top: 35px;
+    left: 85px;
+    font-size: 13px;
+}
+.dashboard-lists-userImages{
+    position: absolute;
+    top: 17px;
+    right: 125px;
+    width: auto;
+}
+.dashboard-lists-userImages-numbers{
+    float: left;
+    width: 28px;
+    margin-left: -15px;
+    cursor: pointer;
+
+}
+.dashboard-lists-userImages:hover .dashboard-lists-userImages-numbers,
+.dashboard-lists-userImages:active .dashboard-lists-userImages-numbers{
+    margin-left: 3px;
+}
+
+.dashboard-lists-comment{
+    position: absolute;
+    top: 13px;
+    right: 30px;
+    font-size: 22px;
+    cursor: pointer;
+}
