+++ web-app/old-files/header.html
@@ -2,13 +2,13 @@
 <html lang="ko">
 <head>
     <meta charset="UTF-8">
-    <link rel="stylesheet" href="../css/header.css">
+    <link rel="stylesheet" href="header.css">
     <script src="https://kit.fontawesome.com/b63e743ce0.js" crossorigin="anonymous"></script>
 </head>
 <body>
 <header class="header">
     <h1 class="header-title">
-        <a href="../main.html">
+        <a href="main.html">
             <img src="../img/logo.png" alt="StudyLog"/>
         </a>
     </h1>
