+++ web-app/old-files/main.html
@@ -4,16 +4,16 @@
     <meta charset="UTF-8">
     <title>studylog</title>
 
-    <link rel="stylesheet" href="css/main.css">
+    <link rel="stylesheet" href="main.css">
 
     <script src="js/global.js" defer></script>
     <script src="https://kit.fontawesome.com/b63e743ce0.js" crossorigin="anonymous"></script>
 </head>
 <body>
-<iframe class="frame-header" src="global/header.html" frameborder="0" scrolling="no"></iframe>
-<iframe class="frame-header-user-info hide" src="global/header_user_info.html"></iframe>
-<iframe class="frame-login hide" src="global/login.html"></iframe>
-<iframe class="frame-sidebar show" src="global/sidebar.html" frameborder="0" scrolling="no"></iframe>
+<iframe class="frame-header" src="header.html" frameborder="0" scrolling="no"></iframe>
+<iframe class="frame-header-user-info hide" src="header_user_info.html"></iframe>
+<iframe class="frame-login hide" src="login.html"></iframe>
+<iframe class="frame-sidebar show" src="sidebar.html" frameborder="0" scrolling="no"></iframe>
 
 <div class="main-container reduce">
     <section class="main-section">
