+++ web-app/old-files/post_list.html
@@ -1,0 +1,60 @@
+<!DOCTYPE html>
+<html lang="ko">
+<head>
+    <meta charset="UTF-8">
+    <title>studylog</title>
+
+    <link rel="stylesheet" href="post_list.css">
+
+    <script src="js/global.js" defer></script>
+    <script src="https://kit.fontawesome.com/b63e743ce0.js" crossorigin="anonymous"></script>
+</head>
+<body>
+<iframe class="frame-header" src="header.html" frameborder="0" scrolling="no"></iframe>
+<iframe class="frame-header-user-info hide" src="header_user_info.html"></iframe>
+<iframe class="frame-login hide" src="login.html"></iframe>
+<iframe class="frame-sidebar show" src="sidebar.html" frameborder="0" scrolling="no"></iframe>
+
+<div class="main-container reduce">
+    <div class="post-list">
+        <h2 class="post-list-title">Web</h2>
+        <span class="post-list-bar"></span>
+        <ul>
+            <li class="post-list-items">
+                <h3 class="post-list-items-title">HTTP</h3>
+                <h4 class="post-list-items-date">2020.12.13</h4>
+                <div class="post-list-items-text">
+                    브라우저 동작 원리
+                    브라우저의 주요 기능
+                    사용자가 참조 하고자 하는 웹페이지를 서버에 요청(Request)하고 응답(Response)을 받아 브라우저에 표시하는 것이다.
+                    요청한 자원은 보통 HTML 문서이지만 PDF나 이미지(image) 또는 다른 형태일 수 도 있다.
+                    자원의 주소는 URI(Uniform Resource Identifier)에 의해 정해진다.
+                </div>
+<!--                <span class="post-list-items-bar post-list-items-barTop"></span>-->
+<!--                <span class="post-list-items-bar post-list-items-barLeft"></span>-->
+<!--                <span class="post-list-items-bar post-list-items-barRight"></span>-->
+<!--                <span class="post-list-items-bar post-list-items-barBottom"></span>-->
+            </li>
+            <li class="post-list-items">
+                <h3 class="post-list-items-title">REST</h3>
+                <h4 class="post-list-items-date">2020.12.12</h4>
+                <div class="post-list-items-text">
+                    REST 란?
+                    "웹에 존재하는 모든 자원(이미지, 동영상, DB 자원)에 고유한 URI를 부여해 활용"
+                    하는 것으로, 자원을 정의하고 자원에 대한 주소를 지정하는 방법론을 의미한다.
+                    Rest의 정의
+                    REST는 REpresentational State Transfer라는 용어의 약자이다.
+                    자원을 이름(자원의 표현)으로 구분하여 해당 자원의 상태(정보)를 주고 받는 모든 것을 의미한다. 즉, 자원(resource)의 표현(representation)에 의한
+                    상태 전달이다.
+                </div>
+<!--                <span class="post-list-items-bar post-list-items-barTop"></span>-->
+<!--                <span class="post-list-items-bar post-list-items-barLeft"></span>-->
+<!--                <span class="post-list-items-bar post-list-items-barRight"></span>-->
+<!--                <span class="post-list-items-bar post-list-items-barBottom"></span>-->
+            </li>
+
+        </ul>
+    </div>
+</div>
+</body>
+</html>
