+++ web-app/old-files/post_write.css
@@ -1,0 +1,131 @@
+@charset "utf-8";
+@import url(reset.css);
+
+.editor-container {
+    margin: 50px 0 50px 5%;
+    width: 90%;
+}
+
+.editor-category {
+    font-size: large;
+    margin-bottom: 10px;
+}
+
+.editor-category-input {
+    font-size: large;
+    vertical-align: top;
+}
+
+.editor-category-input::placeholder {
+    font-size: large;
+}
+
+.editor-category-item {
+    color: var(--main-color);
+    margin-right: 10px
+}
+
+.editor-title {
+    width: 100%;
+    height: auto;
+    margin: 10px 0;
+    font-size: 30px;
+}
+
+.editor-title::placeholder {
+    font-size: 30px;
+}
+
+.editor-buttons {
+    text-align: center;
+}
+
+.save-button {
+    float: right;
+}
+
+.save-button {
+    margin-top: 30px;
+    box-sizing: border-box;
+    padding: 10px 35px;
+    border: 1px solid gray;
+    font-size: larger;
+    cursor: pointer;
+}
+
+.save-button:hover,
+.save-button:active,
+.submit-button:hover,
+.submit-button:active {
+    background: skyblue;
+    color: #ffffff;
+}
+
+.viewer-container {
+    background: rgba(0, 0, 0, 0.7);
+    animation: fadein 0.5s;
+    -moz-animation: fadein 0.5s; /* Firefox */
+    -webkit-animation: fadein 0.5s; /* Safari and Chrome */
+    -o-animation: fadein 0.5s; /* Opera */
+    position: fixed;
+    z-index: 5;
+    width: 100%;
+    height: 100%;
+}
+
+.viewer-container.hide {
+    display: none;
+}
+
+.viewer {
+    padding: 1%;
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    width: 800px;
+    height: 80%;
+    background: whitesmoke;
+    overflow-y: scroll;
+}
+
+.viewer-title {
+    font-size: xxx-large;
+    font-weight: bolder;
+}
+
+.viewer-category {
+    font-size: x-large;
+    color: var(--main-color);
+}
+
+#viewer {
+    margin-top: 20px;
+}
+
+.viewer-buttons {
+    position: absolute;
+    bottom: 0;
+    right: 0;
+    margin-right: 20px;
+    margin-bottom: 20px;
+}
+
+.submit-button, .cancel-button {
+    background-color: white;
+    box-sizing: border-box;
+    padding: 10px 35px;
+    margin-right: 20px;
+    font-size: larger;
+    cursor: pointer;
+}
+
+.cancel-button:hover,
+.cancel-button:active {
+    color: white;
+    background-color: red;
+}
+
+.CodeMirror-vscrollbar {
+    z-index: 1;
+}
