+++ web-app/old-files/post_write.html
@@ -1,0 +1,55 @@
+<!DOCTYPE html>
+<html lang="ko">
+<head>
+    <meta charset="UTF-8">
+    <title>studylog</title>
+
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/codemirror/5.48.4/codemirror.min.css"/>
+    <link rel="stylesheet" href="https://uicdn.toast.com/editor/latest/toastui-editor.min.css"/>
+
+    <link rel="stylesheet" href="main.css">
+    <link rel="stylesheet" href="post_write.css">
+
+    <script src="https://uicdn.toast.com/editor/latest/toastui-editor-all.min.js" defer></script>
+
+    <script src="js/global.js" defer></script>
+    <script src="js/post_write.js" defer></script>
+
+    <script src="https://kit.fontawesome.com/b63e743ce0.js" crossorigin="anonymous"></script>
+</head>
+<body>
+<iframe class="frame-header" src="header.html" frameborder="0" scrolling="no"></iframe>
+<iframe class="frame-header-user-info hide" src="header_user_info.html"></iframe>
+<iframe class="frame-login hide" src="login.html"></iframe>
+<iframe class="frame-sidebar show" src="sidebar.html" frameborder="0" scrolling="no"></iframe>
+
+<article class="viewer-container hide">
+    <div class="viewer">
+        <h1 class="viewer-title"></h1>
+        <h3 class="viewer-category"></h3>
+        <div id="viewer"></div>
+    </div>
+    <div class="viewer-buttons">
+        <button class="submit-button">제출하기</button>
+        <button class="cancel-button">취소</button>
+    </div>
+</article>
+
+<div class="main-container reduce">
+    <div class="editor-container">
+        <input type="text" class="editor-title" placeholder="제목을 입력하세요"/>
+
+        <div class="editor-category">
+            <input class="editor-category-input" type="text" placeholder="카테고리 입력" maxlength="10"/>
+        </div>
+
+        <div id="editorInput" class="editor-input">
+        </div>
+
+        <div id="editor" class="editor"></div>
+
+        <button class="save-button">저장</button>
+    </div>
+</div>
+</body>
+</html>
