--- web-app/postWritePage.html
@@ -1,11 +1,0 @@
-<html>
-<head>
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/codemirror/5.48.4/codemirror.min.css" />
-    <link rel="stylesheet" href="https://uicdn.toast.com/editor/latest/toastui-editor.min.css" />
-</head>
-
-<body>
-<div id="editor"></div>
-<script src="https://uicdn.toast.com/editor/latest/toastui-editor-all.min.js"></script>
-</body>
-</html>
